+++ src/Moeda.java
@@ -1,18 +1,18 @@
 public abstract class Moeda {
-    protected Float valor;
+    protected Double valor;
 
     public Moeda(){}
 
-    public Moeda(Float valor){
+    public Moeda(Double valor){
         this.valor = valor;
     }
 
-    public Float getValor(){
+    public Double getValor(){
         return this.valor;
     }
 
     public abstract void info();
 
     // Método abstrato para ser sobrescrevido
-    public abstract Float converter(Moeda moeda);
+    public abstract Double converter(Moeda moeda);
 }
